Share setupContextMiddleware impl.
