+++ client/boot/common.js
@@ -6,9 +6,7 @@
 
 import debugFactory from 'debug';
 import page from 'page';
-import qs from 'querystring';
 import { some, startsWith } from 'lodash';
-import url from 'url';
 
 /**
  * Internal dependencies
@@ -25,6 +23,7 @@ import { getSections, setupRoutes } from 'sections-middleware';
 import { checkFormHandler } from 'lib/protect-form';
 import notices from 'notices';
 import authController from 'auth/controller';
+import setupContextMiddleware from './page-context-middleware';
 
 const debug = debugFactory( 'calypso' );
 
@@ -34,47 +33,6 @@ const switchUserLocale = ( currentUser, reduxStore ) => {
 	if ( localeSlug ) {
 		reduxStore.dispatch( setLocale( localeSlug ) );
 	}
-};
-
-const setupContextMiddleware = reduxStore => {
-	page( '*', ( context, next ) => {
-		// page.js url parsing is broken so we had to disable it with `decodeURLComponents: false`
-		const parsed = url.parse( context.canonicalPath, true );
-		context.prevPath = parsed.path === context.path ? false : parsed.path;
-		context.query = parsed.query;
-
-		context.hashstring = ( parsed.hash && parsed.hash.substring( 1 ) ) || '';
-		// set `context.hash` (we have to parse manually)
-		if ( context.hashstring ) {
-			try {
-				context.hash = qs.parse( context.hashstring );
-			} catch ( e ) {
-				debug( 'failed to query-string parse `location.hash`', e );
-				context.hash = {};
-			}
-		} else {
-			context.hash = {};
-		}
-
-		context.store = reduxStore;
-
-		// client version of the isomorphic method for redirecting to another page
-		context.redirect = ( httpCode, newUrl = null ) => {
-			if ( isNaN( httpCode ) && ! newUrl ) {
-				newUrl = httpCode;
-			}
-
-			return page.replace( newUrl, context.state, false, false );
-		};
-
-		// Break routing and do full load for logout link in /me
-		if ( context.pathname === '/wp-login.php' ) {
-			window.location.href = context.path;
-			return;
-		}
-
-		next();
-	} );
 };
 
 // We need to require sections to load React with i18n mixin
