+++ client/boot/page-context-middleware.js
@@ -1,0 +1,52 @@
+/**
+ * External dependencies
+ */
+import debugFactory from 'debug';
+import page from 'page';
+import qs from 'querystring';
+import url from 'url';
+
+const debug = debugFactory( 'calypso' );
+
+const setupContextMiddleware = reduxStore => {
+	page( '*', ( context, next ) => {
+		// page.js url parsing is broken so we had to disable it with `decodeURLComponents: false`
+		const parsed = url.parse( context.canonicalPath, true );
+		context.prevPath = parsed.path === context.path ? false : parsed.path;
+		context.query = parsed.query;
+
+		context.hashstring = ( parsed.hash && parsed.hash.substring( 1 ) ) || '';
+		// set `context.hash` (we have to parse manually)
+		if ( context.hashstring ) {
+			try {
+				context.hash = qs.parse( context.hashstring );
+			} catch ( e ) {
+				debug( 'failed to query-string parse `location.hash`', e );
+				context.hash = {};
+			}
+		} else {
+			context.hash = {};
+		}
+
+		context.store = reduxStore;
+
+		// client version of the isomorphic method for redirecting to another page
+		context.redirect = ( httpCode, newUrl = null ) => {
+			if ( isNaN( httpCode ) && ! newUrl ) {
+				newUrl = httpCode;
+			}
+
+			return page.replace( newUrl, context.state, false, false );
+		};
+
+		// Break routing and do full load for logout link in /me
+		if ( context.pathname === '/wp-login.php' ) {
+			window.location.href = context.path;
+			return;
+		}
+
+		next();
+	} );
+};
+
+export default setupContextMiddleware;
